Reorganize files and rename main entrty point
Attempting to make all of the various ways that the script can be run
happy.
