+++ mbget/barrel_asset.py
@@ -1,4 +1,4 @@
-from impl.version import Version
+from mbget.version import Version
 
 class BarrelAsset(object):
     def __init__(self, asset_name: str, asset_version: str, asset_content: bytes):
